splitting into 2 projects
