+++ ApplicationIJP/src/main/java/com/ukgG3/Application/IjpApplication.java
@@ -1,0 +1,13 @@
+package com.ukgG3.Application;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class IjpApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(IjpApplication.class, args);
+	}
+
+}
