+++ ApplicationIJP/src/main/java/com/ukgG3/Application/entity/ApplicationStatus.java
