+++ ApplicationIJP/src/main/resources/application.yml
@@ -1,0 +1,25 @@
+server:
+  port: 8081
+spring:
+  application:
+    name: "application"
+  devtools:
+    livereload:
+      enabled: true
+    restart:
+      enabled: true
+  datasource:
+    url: jdbc:h2:mem:testdb
+    driverClassName: org.h2.Driver
+    username: sa
+    password: ''
+  jpa:
+    database-platform: org.hibernate.dialect.H2Dialect
+    hibernate:
+      ddl-auto: update
+  h2:
+    console:
+      enabled: true
+  sql:
+    init:
+      mode: always
