+++ ApplicationIJP/src/test/java/com/ukgG3/Application/IjpApplicationTests.java
@@ -1,0 +1,13 @@
+package com.ukgG3.Application;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class IjpApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
