--- IJP/src/main/java/com/ukgG3/JobPosting/IjpApplication.java
@@ -1,13 +1,0 @@
-package com.ukgG3.JobPosting;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class IjpApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(IjpApplication.class, args);
-	}
-
-}
