+++ JobPostIJP/src/main/java/com/ukgG3/JobPosting/JobPosting.java
@@ -1,0 +1,11 @@
+package com.ukgG3.JobPosting;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class JobPosting {
+	public static void main(String[] args) {
+		SpringApplication.run(JobPosting.class, args);
+	}
+}
