+++ JobPostIJP/src/main/java/com/ukgG3/JobPosting/dto/EmployeeDto.java
@@ -1,0 +1,17 @@
+package com.ukgG3.JobPosting.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.time.LocalDate;
+
+@Getter @Setter
+public class EmployeeDto {
+    private long employee_id;
+    private String first_name;
+    private String last_name;
+    private String email;
+    private String password;
+    private LocalDate dob;
+    private String role;
+}
